+++ src/zope/tal/talgenerator.py
@@ -15,14 +15,7 @@
 Code generator for :class:`~.TALInterpreter` intermediate code.
 """
 import re
-
-
-try:
-    # Python 3.x
-    from html import escape
-except ImportError:
-    # Python 2.x
-    from cgi import escape
+from html import escape
 
 from zope.tal import taldefs
 from zope.tal.taldefs import NAME_RE
@@ -35,16 +28,10 @@ from zope.tal.translationcontext import DEFAULT_DOMAIN
 from zope.tal.translationcontext import TranslationContext
 
 
-try:
-    xrange
-except NameError:
-    xrange = range  # Python 3.x
-
-
 _name_rx = re.compile(NAME_RE)
 
 
-class TALGenerator(object):
+class TALGenerator:
     """
     Generate intermediate code.
     """
@@ -96,7 +83,7 @@ class TALGenerator(object):
         output = []
         collect = []
         cursor = 0
-        for cursor in xrange(len(program) + 1):
+        for cursor in range(len(program) + 1):
             try:
                 item = program[cursor]
             except IndexError:
@@ -169,7 +156,7 @@ class TALGenerator(object):
     def optimizeStartTag(self, collect, name, attrlist, end):
         # return true if the tag can be converted to plain text
         if not attrlist:
-            collect.append("<%s%s" % (name, end))
+            collect.append("<{}{}".format(name, end))
             return 1
         opt = 1
         new = ["<" + name]
@@ -183,7 +170,7 @@ class TALGenerator(object):
                 if item[1] is None:
                     s = item[0]
                 else:
-                    s = '%s="%s"' % (item[0], taldefs.attrEscape(item[1]))
+                    s = '{}="{}"'.format(item[0], taldefs.attrEscape(item[1]))
                 attrlist[i] = item[0], s
                 new.append(" " + s)
         # if no non-optimizable attributes were found, convert to plain text
@@ -225,8 +212,11 @@ class TALGenerator(object):
         try:
             return self.expressionCompiler.compile(expr)
         except self.CompilerError as err:
-            raise TALError('%s in expression %s' % (err.args[0], repr(expr)),
-                           self.position)
+            raise TALError(
+                '{} in expression {}'.format(
+                    err.args[0],
+                    repr(expr)),
+                self.position)
 
     def pushProgram(self):
         self.stack.append(self.program)