+++ src/zope/tal/tests/test_sourcepos.py
@@ -14,14 +14,7 @@
 """Tests for TALInterpreter.
 """
 import unittest
-
-
-try:
-    # Python 2.x
-    from StringIO import StringIO
-except ImportError:
-    # Python 3.x
-    from io import StringIO
+from io import StringIO
 
 from zope.tal.dummyengine import DummyEngine
 from zope.tal.htmltalparser import HTMLTALParser
@@ -90,4 +83,4 @@ class SourcePosTestCase(unittest.TestCase):
 
 
 def test_suite():
-    return unittest.makeSuite(SourcePosTestCase)
+    return unittest.defaultTestLoader.loadTestsFromTestCase(SourcePosTestCase)